power/requirements.txt
# Core dependencies
google-generativeai>=0.8.0
pydantic>=2.0.0
typing-extensions>=4.0.0

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0

# Code quality
pylint>=3.0.0
black>=23.0.0
isort>=5.0.0
mypy>=1.0.0

# Utilities
python-dotenv>=1.0.0